Add support for printing UTF16, UTF32 and wchar into UTF8 strings Add UTF helper library to decode, encode, copy and query UTF strings

+++ MSF_FormatPrint.h
@@ -32,7 +32,7 @@ struct MSF_PrintData
 	size_t myMaxLength; // max length required, can be an estimate but must never be too low.
 
 	// store any data you may need later
-	mutable size_t myUserData;
+	mutable uint64_t myUserData;
 
 	// %<width>.<precision>
 	uint16_t myWidth;
